requirements.txt
django==4.1.7
djangorestframework
django-filter
django-environ
environ
